Moving code due to HACS instaal problems
