--- card/do_not_delete.md
@@ -1,1 +1,0 @@
-do not delete
