+++ src/esc-awning-bottom.png
